feat: Upgrade dependencies and runtime to latest

+++ src/main/resources/db/migration/V1__create_basic_schema.sql
@@ -1,20 +1,20 @@
 CREATE TABLE categories (
-    id INTEGER PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     name VARCHAR NOT NULL,
-    parent_id INTEGER REFERENCES categories (id)
+    parent_id BIGINT REFERENCES categories (id)
 );
 
 CREATE TABLE items (
-    id INTEGER PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     name VARCHAR NOT NULL,
     default_cost NUMERIC(12,2),
-    category_id INTEGER REFERENCES categories (id)
+    category_id BIGINT REFERENCES categories (id)
 );
 
 CREATE TABLE expenses(
-    id INTEGER PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
-    item_id INTEGER REFERENCES items (id),
-    category_id INTEGER REFERENCES categories (id),
+    id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    item_id BIGINT REFERENCES items (id),
+    category_id BIGINT REFERENCES categories (id),
     amount NUMERIC(12,2) NOT NULL,
     description TEXT,
     item_count INTEGER,
@@ -23,9 +23,9 @@ CREATE TABLE expenses(
 );
 
 CREATE TABLE incomes(
-    id INTEGER PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     amount NUMERIC(12,2) NOT NULL,
-    category_id INTEGER REFERENCES categories(id),
+    category_id BIGINT REFERENCES categories(id),
     description VARCHAR,
     received_at TIMESTAMPTZ NOT NULL,
     created_at TIMESTAMPTZ NOT NULL
