+++ src/main/resources/db/migration/V2__add_expense_lists.sql
@@ -1,8 +1,8 @@
 CREATE TABLE expense_lists(
-    id INTEGER PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     list_date DATE NOT NULL,
     total DECIMAL NOT NULL
 );
 
 ALTER TABLE expenses
-    ADD COLUMN list_id INTEGER REFERENCES expense_lists (id);
+    ADD COLUMN list_id BIGINT REFERENCES expense_lists (id);
