+++ src/main/resources/db/migration/V3__add_members.sql
@@ -1,20 +1,20 @@
 CREATE TABLE members(
-    id INTEGER PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     name VARCHAR
 );
 
 ALTER TABLE expenses
-    ADD COLUMN member_id INTEGER REFERENCES members (id);
+    ADD COLUMN member_id BIGINT REFERENCES members (id);
 
 ALTER TABLE incomes
-    ADD COLUMN member_id INTEGER REFERENCES members(id);
+    ADD COLUMN member_id BIGINT REFERENCES members(id);
 
 CREATE TABLE groups(
-    id INTEGER PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     name varchar
 );
 
 CREATE TABLE group_members(
-    group_id INTEGER REFERENCES groups (id),
-    member_id INTEGER REFERENCES members (id)
+    group_id BIGINT REFERENCES groups (id),
+    member_id BIGINT REFERENCES members (id)
 );
