+++ src/main/resources/db/migration/V4__add_group_to_expenses.sql
@@ -1,2 +1,2 @@
 ALTER TABLE expenses
-    ADD COLUMN group_id INTEGER REFERENCES groups (id);
+    ADD COLUMN group_id BIGINT REFERENCES groups (id);
