+++ src/main/resources/db/migration/V5__add_category_rules.sql
@@ -1,8 +1,8 @@
 CREATE TABLE category_rules(
-    id INTEGER PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
     description_rule VARCHAR,
     category_rule VARCHAR,
     description_replacement VARCHAR,
     importance INTEGER DEFAULT 1,
-    category_id INTEGER REFERENCES categories(id)
+    category_id BIGINT REFERENCES categories(id)
 );
